chore(project): Reorganize project structure
Move all APIs under `/sys` to lib/sys/ directory, and `/secret` to
the `lib/backends/` to streamline and organize the code base to
deal with the fact the code base will very likely continue to grow.
