+++ lib/sys/auth.js
