+++ lib/sys/init.js
