+++ lib/sys/keys.js
@@ -2,7 +2,7 @@ var
   Vaulted = {},
   Promise = require('bluebird'),
   _ = require('lodash'),
-  internal = require('./internal');
+  internal = require('../internal');
 
 /**
   * @module keys
