+++ lib/sys/leader.js
