+++ lib/sys/mounts.js
@@ -2,7 +2,7 @@ var
   Vaulted = {},
   Promise = require('bluebird'),
   _ = require('lodash'),
-  utils = require('./utils');
+  utils = require('../utils');
 
 /**
  * @module mounts
