+++ lib/sys/seal.js
