+++ lib/vaulted.js
@@ -146,23 +146,23 @@ Vaulted.prototype.validateEndpoint = function validateEndpoint(endpoint, mountNa
 
 
 // sys
-require('./auth')(Vaulted.prototype);
-require('./audit')(Vaulted.prototype);
-require('./health')(Vaulted.prototype);
-require('./init')(Vaulted.prototype);
-require('./keys')(Vaulted.prototype);
-require('./leader')(Vaulted.prototype);
-require('./leases')(Vaulted.prototype);
-require('./mounts')(Vaulted.prototype);
-require('./policy')(Vaulted.prototype);
-require('./seal')(Vaulted.prototype);
+require('./sys/auth')(Vaulted.prototype);
+require('./sys/audit')(Vaulted.prototype);
+require('./sys/health')(Vaulted.prototype);
+require('./sys/init')(Vaulted.prototype);
+require('./sys/keys')(Vaulted.prototype);
+require('./sys/leader')(Vaulted.prototype);
+require('./sys/leases')(Vaulted.prototype);
+require('./sys/mounts')(Vaulted.prototype);
+require('./sys/policy')(Vaulted.prototype);
+require('./sys/seal')(Vaulted.prototype);
 
 // auth backends
 require('./auth/token')(Vaulted.prototype);
 require('./auth/appid')(Vaulted.prototype);
 
 // secrets backends
-require('./secret')(Vaulted.prototype);
+require('./backends/secret')(Vaulted.prototype);
 require('./backends/cubbyhole')(Vaulted.prototype);
 require('./backends/consul')(Vaulted.prototype);
 require('./backends/pki')(Vaulted.prototype);
